requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
PyJWT